requirements.txt
pyyaml==6.0.1
flickrapi==2.4.0
